@@ -2,74 +2,18 @@
 
 namespace App\Entity;
 
-use App\Repository\ItemRepository;
-use Doctrine\ORM\Mapping as ORM;
+use DateTime;
 
-/**
- * @ORM\Entity(repositoryClass=ItemRepository::class)
- */
 class Item
 {
-    /**
-     * @ORM\Id
-     * @ORM\GeneratedValue
-     * @ORM\Column(type="integer")
-     */
-    private $id;
+    public ?string $name;
+    public ?string $content;
+    public DateTime $dateCreation;
 
-    /**
-     * @ORM\Column(type="string", length=255)
-     */
-    private $name;
-
-    /**
-     * @ORM\Column(type="string", length=1000, nullable=true)
-     */
-    private $content;
-
-    /**
-     * @ORM\Column(type="date")
-     */
-    private $dateCreation;
-
-    public function getId(): ?int
-    {
-        return $this->id;
-    }
-
-    public function getName(): ?string
-    {
-        return $this->name;
-    }
-
-    public function setName(string $name): self
+    public function __construct(?string $name, ?string $content)
     {
         $this->name = $name;
-
-        return $this;
-    }
-
-    public function getContent(): ?string
-    {
-        return $this->content;
-    }
-
-    public function setContent(?string $content): self
-    {
         $this->content = $content;
-
-        return $this;
-    }
-
-    public function getDateCreation(): ?\DateTimeInterface
-    {
-        return $this->dateCreation;
-    }
-
-    public function setDateCreation(\DateTimeInterface $dateCreation): self
-    {
-        $this->dateCreation = $dateCreation;
-
-        return $this;
+        $this->dateCreation = new DateTime();
     }
 }
