@@ -3,11 +3,15 @@ declare(strict_types=1);
 
 use PHPUnit\Framework\TestCase;
 use App\Entity\User;
+use App\Services\ToDoList as ToDoListService;
+use App\Entity\ToDoList as ToDoListEntity;
+use App\Services\ArrayUtils;
 
 class UserTest extends TestCase {
+
     protected function setUp(): void
     {
-        parent::setUp(); // TODO: Change the autogenerated stub
+        parent::setUp();
     }
 
     public function testIsValidWithGoodInfos()
@@ -16,7 +20,9 @@ class UserTest extends TestCase {
             "louispelarrey@gmail.com",
             "Pelarrey",
             "Louis",
-            date("Y-m-d", strtotime('-13 years'))
+            date("Y-m-d", strtotime('-13 years')),
+            "azerty",
+            new ToDoListService(new ToDoListEntity(), new ArrayUtils())
         );
         $this->assertTrue($user->isValid());
     }
@@ -27,7 +33,9 @@ class UserTest extends TestCase {
             "uuu",
             "Pelarrey",
             "Louis",
-            "2000-05-21"
+            "2000-05-21",
+            "azerty",
+            new ToDoListService(new ToDoListEntity(), new ArrayUtils())
         );
         $this->assertFalse($user->isValid());
     }
@@ -38,7 +46,9 @@ class UserTest extends TestCase {
             "louispelarrey@gmail.com",
             null,
             "Louis",
-            "2000-05-21"
+            "2000-05-21",
+            "azerty",
+            new ToDoListService(new ToDoListEntity(), new ArrayUtils())
         );
         $this->assertFalse($user->isValid());
     }
@@ -49,7 +59,9 @@ class UserTest extends TestCase {
             "louispelarrey@gmail.com",
             "Pelarrey",
             null,
-            "2000-05-21"
+            "2000-05-21",
+            "azerty",
+            new ToDoListService(new ToDoListEntity(), new ArrayUtils())
         );
         $this->assertFalse($user->isValid());
     }
@@ -60,7 +72,9 @@ class UserTest extends TestCase {
             "louispelarrey@gmail.com",
             "Pelarrey",
             "Louis",
-            "2018-05-21"
+            "2018-05-21",
+            "azerty",
+            new ToDoListService(new ToDoListEntity(), new ArrayUtils())
         );
         $this->assertFalse($user->isValid());
     }
@@ -71,7 +85,9 @@ class UserTest extends TestCase {
             "louispelarrey@gmail.com",
             "",
             "Louis",
-            "2000-05-21"
+            "2000-05-21",
+            "azerty",
+            new ToDoListService(new ToDoListEntity(), new ArrayUtils())
         );
         $this->assertFalse($user->isValid());
     }
@@ -82,7 +98,22 @@ class UserTest extends TestCase {
             "louispelarrey@gmail.com",
             "Pelarrey",
             "Louis",
-            null
+            null,
+            "azerty",
+            new ToDoListService(new ToDoListEntity(), new ArrayUtils())
+        );
+        $this->assertFalse($user->isValid());
+    }
+
+    public function testIsNotWorkingWithNothingPassword()
+    {
+        $user = new User(
+            "louispelarrey@gmail.com",
+            "Pelarrey",
+            "Louis",
+            date("Y-m-d", strtotime('-13 years')),
+            "",
+            new ToDoListService(new ToDoListEntity(), new ArrayUtils())
         );
         $this->assertFalse($user->isValid());
     }
